Update README.md
I just got this working on heroku. There is a bit of chicken-and-egg problem with the force_hostname field because visiting /admin when you are not an administrator brings you nowhere. Was there perhaps an additional seed data step that was not included in this read me? Also - the assets precompile will always fail the first time until you have a database setup. Not sure how you want to handle this in this project.

Overall the setup was pretty smooth! Thanks for the contribution!